pyproject.toml
[project]
name = "clip-dlut"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14.2"
dependencies = [
    "celery>=5.6.2",
    "fastapi>=0.128.3",
    "ftfy>=6.3.1",
    "pillow>=12.1.0",
    "pydantic-settings>=2.12.0",
    "python-multipart>=0.0.22",
    "redis>=7.1.0",
    "regex>=2026.1.15",
    "requests>=2.32.5",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "tqdm>=4.67.1",
    "transformers>=5.0.0",
    "uvicorn>=0.40.0",
]
